admin 폴더 내 service, dao 추가 + 컨트롤러 코드 추가 -admin/blindBrd.do 관련 코드 추가
입양 후기, 자유게시판 수정
-코드 수정
-view.jsp의 '블라인드'버튼 <a> -> <button>으로 변경하고 script 추가
<!-- tpNoStr == tp와 no을 같이 보냄 : 해당 글을 블라인드하도록 -->
						<button id="blindBtn" class="btn btn-danger float-right ml-1
mr-1">블라인드</button>

+++ semiProject/src/com/kh/semiProject/review/model/service/ReviewService.java
@@ -199,22 +199,26 @@ public class ReviewService {
 			
 			if(!iList.isEmpty()) {
 				for(Image img : iList) {
-					
-					//현재 썸머 노트를 통해 저장된 이미지의 주소는 ../resource/uploadImages 
-					//(썸머노트에서 src에 하려면 이 주소만 되어서 상대주소로 작성..)
-					//→ 이미지 삭제를 위해 필요한 주소는 C:/workspace/semi/Chooru_Pj\semiProject\WebContent\resources/uploadImages
-					//→ 바꿔줘야 함
-					String filePath = (String)map.get("root");
-					filePath += img.getFilePath().substring(3);
-					
-					String fileName = img.getFileName();
-					
-					File deleteFile = new File(filePath + fileName);
-					//해당 파일의 전체 주소 : filePath + fileName
-					//File 객체는 전체 주소에 있는 파일 객체를 선택할 때 사용함
-					// → filePath+fileName인 파일이 없다면 deleteFile은 null 값.
-					
-					if(deleteFile.exists()) deleteFile.delete();
+						
+					if(!img.getFilePath().contains("http")) {//복붙하거나 인터넷에서 주소로 가져온 이미지일 때
+						
+						//현재 썸머 노트를 통해 저장된 이미지의 주소는 ../resource/uploadImages 
+						//(썸머노트에서 src에 하려면 이 주소만 되어서 상대주소로 작성..)
+						//→ 이미지 삭제를 위해 필요한 주소는 C:/workspace/semi/Chooru_Pj\semiProject\WebContent\resources/uploadImages
+						//→ 바꿔줘야 함
+						String filePath = (String)map.get("root");
+						
+						filePath += img.getFilePath().substring(3);
+						
+						String fileName = img.getFileName();
+						
+						File deleteFile = new File(filePath + fileName);
+						//해당 파일의 전체 주소 : filePath + fileName
+						//File 객체는 전체 주소에 있는 파일 객체를 선택할 때 사용함
+						// → filePath+fileName인 파일이 없다면 deleteFile은 null 값.
+						
+						if(deleteFile.exists()) deleteFile.delete();
+					}
 				}
 			}
 			
@@ -369,22 +373,25 @@ public class ReviewService {
 			if(!iList.isEmpty()) {
 				for(Image img : iList) {
 					
-					//현재 썸머 노트를 통해 저장된 이미지의 주소는 ../resource/uploadImages 
-					//(썸머노트에서 src에 하려면 이 주소만 되어서 상대주소로 작성..)
-					//→ 이미지 삭제를 위해 필요한 주소는 C:\workspace\semi\Chooru_Pj\semiProject\WebContent\resources/uploadImages
-					//→ 바꿔줘야 함
-					String filePath = (String)map.get("root");
-					filePath += img.getFilePath().substring(3);
+					if(!img.getFilePath().contains("http")) {
 					
-					String fileName = img.getFileName();
-					
-					File deleteFile = new File(filePath + fileName);
-					//해당 파일의 전체 주소 : filePath + fileName
-					//File 객체는 전체 주소에 있는 파일 객체를 선택할 때 사용함
-					// → filePath+fileName인 파일이 없다면 deleteFile은 null 값.
-					
-					if(deleteFile.exists()) {
-						deleteFile.delete();
+						//현재 썸머 노트를 통해 저장된 이미지의 주소는 ../resource/uploadImages 
+						//(썸머노트에서 src에 하려면 이 주소만 되어서 상대주소로 작성..)
+						//→ 이미지 삭제를 위해 필요한 주소는 C:\workspace\semi\Chooru_Pj\semiProject\WebContent\resources/uploadImages
+						//→ 바꿔줘야 함
+						String filePath = (String)map.get("root");
+						filePath += img.getFilePath().substring(3);
+						
+						String fileName = img.getFileName();
+						
+						File deleteFile = new File(filePath + fileName);
+						//해당 파일의 전체 주소 : filePath + fileName
+						//File 객체는 전체 주소에 있는 파일 객체를 선택할 때 사용함
+						// → filePath+fileName인 파일이 없다면 deleteFile은 null 값.
+						
+						if(deleteFile.exists()) {
+							deleteFile.delete();
+						}
 					}
 				}
 			}
@@ -404,21 +411,23 @@ public class ReviewService {
 				
 				for(Image img : deleteImages) {
 					
-					//현재 썸머 노트를 통해 저장된 이미지의 주소는 ../resource/uploadImages 
-					//(썸머노트에서 src에 하려면 이 주소만 되어서 상대주소로 작성..)
-					//→ 이미지 삭제를 위해 필요한 주소는 C:\workspace\semi\Chooru_Pj\semiProject\WebContent\resources/uploadImages
-					//→ 바꿔줘야 함
-					String filePath = (String)map.get("root");
-					filePath += img.getFilePath().substring(3);
-					
-					String fileName = img.getFileName();
-					
-					File deleteFile = new File(filePath + fileName);
-					
-					System.out.println(deleteFile);
-					
-					if(deleteFile.exists()) {
-						deleteFile.delete();
+					if(!img.getFilePath().contains("http")) {
+						//현재 썸머 노트를 통해 저장된 이미지의 주소는 ../resource/uploadImages 
+						//(썸머노트에서 src에 하려면 이 주소만 되어서 상대주소로 작성..)
+						//→ 이미지 삭제를 위해 필요한 주소는 C:\workspace\semi\Chooru_Pj\semiProject\WebContent\resources/uploadImages
+						//→ 바꿔줘야 함
+						String filePath = (String)map.get("root");
+						filePath += img.getFilePath().substring(3);
+						
+						String fileName = img.getFileName();
+						
+						File deleteFile = new File(filePath + fileName);
+						
+						System.out.println(deleteFile);
+						
+						if(deleteFile.exists()) {
+							deleteFile.delete();
+						}
 					}
 				}
 			}
